+++ src/Requests/Excludes.php
@@ -4,27 +4,25 @@ namespace Giadc\JsonApiRequest\Requests;
 
 use Symfony\Component\HttpFoundation\Request;
 
-class Filters implements RequestInterface
+class Excludes implements RequestInterface
 {
-    private $container = array();
+    private array $container = [];
 
-    public function __construct($filters = [])
+    public function __construct(array $excludes = [])
     {
-        $filters = $filters ?: [];
-
-        foreach ($filters as $key => $filter) {
-            $this->add($key, $filter);
+        foreach ($excludes as $key => $exclude) {
+            $this->add($key, $exclude);
         }
     }
 
     public static function fromRequest(Request $request): self
     {
-        $filters = $request->query->get('filter');
-        return new self($filters);
+        $excludes = $request->query->get('excludes');
+        return new self($excludes ?: []);
     }
 
     /**
-     * Add an additional filter
+     * Add an additional exclude
      *
      * @param string|array $value
      */
@@ -38,25 +36,29 @@ class Filters implements RequestInterface
     }
 
     /**
-     * Convert Filters to a string
+     * Convert Excludes to a string
+     *
+     * @return string
      */
     public function toString(): string
     {
         $params = '';
 
-        foreach ($this->container as $key => $filter) {
+        foreach ($this->container as $key => $exclude) {
             if (!$params == '') {
                 $params .= '&';
             }
 
-            $params .= "filter[$key]=" . implode(',', $filter);
+            $params .= "excludes[$key]=" . implode(',', $exclude);
         };
 
         return $params;
     }
 
     /**
-     * Get Filters as a params array
+     * Get Excludes as a params array
+     *
+     * @return array[string]
      */
     public function getParamsArray(): array
     {
@@ -64,7 +66,7 @@ class Filters implements RequestInterface
     }
 
     /**
-     * Get Filters as query string fragment
+     * Get Excludes as query string fragment
      */
     public function getQueryString(): string
     {
@@ -72,14 +74,10 @@ class Filters implements RequestInterface
     }
 
     /**
-     * Get Filters as a multi-dimensional array
+     * Get Excludes as a multi-dimensional array
      */
     public function toArray(): array
     {
-        if (is_null($this->container)) {
-            return [];
-        }
-
         return $this->container;
     }
 }