+++ tests/unit/Requests/SortingTest.php
@@ -12,7 +12,7 @@ class SortingTest extends TestCase
         $this->sorting = new Sorting('width,-height');
     }
 
-    public function test_it_returns_an_array_for_fields_to_sort_by()
+    public function test_it_returns_an_array_for_fields_to_sort_by(): void
     {
         $expected = [
             ['field' => 'width', 'direction' => 'ASC'],
@@ -21,7 +21,7 @@ class SortingTest extends TestCase
         $this->assertEquals($expected, $this->sorting->toArray());
     }
 
-    public function test_it_returns_a_params_array()
+    public function test_it_returns_a_params_array(): void
     {
         $expected = [
             'sort' => 'width,-height',
@@ -30,16 +30,30 @@ class SortingTest extends TestCase
         $this->assertEquals($expected, $this->sorting->getParamsArray());
     }
 
-    public function test_it_returns_a_query_string()
+    public function test_it_returns_a_query_string(): void
     {
         $expected = 'sort=width,-height';
         $this->assertEquals($expected, $this->sorting->getQueryString());
     }
 
-    public function test_it_replaces_sorting()
+    public function test_it_replaces_sorting(): void
     {
         $expected = 'sort=weight,volume';
-        $this->sorting->setSorting('weight,volume');
+        $this->sorting->setWithString('weight,volume');
         $this->assertEquals($expected, $this->sorting->getQueryString());
+    }
+
+    public function test_it_can_add_sorting_field(): void
+    {
+        $expected = [
+            ['field' => 'width', 'direction' => 'ASC'],
+            ['field' => 'name', 'direction' => 'DESC'],
+            ['field' => 'height', 'direction' => 'ASC'],
+        ];
+
+        $this->sorting->add('name', 'desc');
+        $this->sorting->add('height', 'ASC');
+
+        $this->assertEqualsCanonicalizing($expected, $this->sorting->toArray());
     }
 }