+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- Added new options, `fields` and `excludes`.
+
+### Changed
+- RequestParams now allows for the modification of fields.
+
 ## [1.2.0] - 2019-12-09
 ### Changed
 - Lessened restrictions on dependency versions.